[NP] Assignment 1 Problem 2 Bug Fix - Avoid buffered streams
At the client side we know we will get each message as 5 bytes always,
so reduce the buffer size to 5 bytes and always read the same. Having
bigger buffer could result in reading multiple messages on the TCP
socket as that is sliding window protocol and is a bytes stream and club
two message bytes together. This will require as to have message
boundaries to parse the messages correctly. Insteam at on instance read
only 5 bytes so that we get only one message per read.

Also update some log messages to have marker for hello message and
revert the message count before server send "hello" to 20 as asked in
the problem

Signed-off-by: Shobhit Kumar <kumar@shobhit.info>